Make course a Definition::ValueObject
