+++ .gitignore
@@ -9,7 +9,7 @@
 
 # Ignore all logfiles and tempfiles.
 /log/*
-log/*.log
+log/*
 tmp/*
 /tmp/*
 !/log/.keep
